adding the framework for the front end of the game itself
adding a frame to the MainWindow.xaml so that we have the ability to
quickly change the layout of the application.
creating helpful window to hold the help page for us and have it easily
accessed.
creating a main page object to hold the layout details for the main page
page.
creating a help page object to hold the layout details for the help
page.
creating a game page object to hold the layout details for the game
page.
creating an IPageChanger Interface to allow our pages to communicate
effectively with the mainwindow

creating a PageRequest enum to hold all the pages that we will be able
to navigate too.
adding an html file to display the rules of the game taken from the
pente game website.
adding a test.png file in order to test out some elements of the user

creating a player control to centralize the look and feel for player UI.
cre